+++ activesupport/test/core_ext/string_ext_test.rb
@@ -836,10 +836,9 @@ class OutputSafetyTest < ActiveSupport::TestCase
 
   test "Adding an object not responding to `#to_str` to a safe string is deprecated" do
     string = @string.html_safe
-    assert_deprecated("Implicit conversion of #{@to_s_object.class} into String by ActiveSupport::SafeBuffer is deprecated", ActiveSupport.deprecator) do
+    assert_raises(NoMethodError) do
       string << @to_s_object
     end
-    assert_equal "helloto_s", string
   end
 
   test "Adding an object to a safe string returns a safe string" do